requirements.txt
customtkinter==5.2.1
pillow==10.1.0
psutil==5.9.6
